Tweaks for multipft ensemble runs

+++ rscripts/write.configs.R
@@ -5,7 +5,7 @@ if(interactive()){
   if(user == 'ed'){
     settings.file = '~/in/ebifarm/fast/ebifarm.pavi.xml'
   } else if(user == 'davids14') {
-    settings.file = '~/pecan/tundra.xml'
+    settings.file = '~/pecan/atqasuk.xml'
   } else {
     paste('please specify settings file in write.configs.R')
   }
@@ -26,43 +26,46 @@ outdirs <- unlist(xpathApply(settings.xml, '//pfts//pft//outdir', xmlValue))
 trait.samples <- list()
 sa.samples <- list()
 ensemble.samples <- list()
+env.samples <- list()
 
 ## Remove existing config files
-
-todelete <- dir(paste(settings$pfts$pft$outdir, 'out/', sep = ''),
-                c('ED2INc.*','c.*'),
-                recursive=TRUE, full.names = TRUE)
-if(length(todelete>0)) file.remove(todelete)
-
-filename.root <- get.run.id('c.','ebifarm.pavi')
-
-if(host$name == 'localhost'){
-  if(length(dir(host$rundir, pattern = filename.root)) > 0) {
-    todelete <- dir(host$outdir,
-                    pattern = paste(filename.root, "*[^log]", sep = ''), 
-                    recursive=TRUE, full.names = TRUE)
-    file.remove(todelete)
-  }
-} else {
-  files <- system(paste("ssh ", host$name, " 'ls ", host$rundir, "*", filename.root, "*'", sep = ''), intern = TRUE)
-  if(length(files) > 0 ) {
-    todelete <- files[-grep('log', files)]
-    system(paste("ssh -T ", host$name,
-                 " 'for f in ", paste(todelete, collapse = ' '),"; do rm $f; done'",sep=''))
+if(FALSE){
+  todelete <- dir(paste(settings$pfts$pft$outdir, 'out/', sep = ''),
+      c('ED2INc.*','c.*'),
+      recursive=TRUE, full.names = TRUE)
+  if(length(todelete>0)) file.remove(todelete)
+  
+  filename.root <- get.run.id('c.','*')
+  
+  if(host$name == 'localhost'){
+    if(length(dir(host$rundir, pattern = filename.root)) > 0) {
+      todelete <- dir(host$outdir,
+          pattern = paste(filename.root, "*[^log]", sep = ''), 
+          recursive=TRUE, full.names = TRUE)
+      file.remove(todelete)
+    }
+  } else {
+    files <- system(paste("ssh ", host$name, " 'ls ", host$rundir, "*", filename.root, "*'", sep = ''), intern = TRUE)
+    if(length(files) > 0 ) {
+      todelete <- files[-grep('log', files)]
+      system(paste("ssh -T ", host$name,
+              " 'for f in ", paste(todelete, collapse = ' '),"; do rm $f; done'",sep=''))
+    }
   }
 }
 
-## Load priors and posteriors
+## Load PFT priors and posteriors
 
 for (i in seq(pft.names)){
+  
   load(paste(outdirs[i], 'prior.distns.Rdata', sep=''))
-
+  
   if("trait.mcmc.Rdata" %in% dir(outdirs)) {
     load(paste(outdirs[i], 'trait.mcmc.Rdata', sep=''))
   }
-
+  
   pft.name <- pft.names[i]
-
+  
   ## when no ma for a trait, sample from  prior
   ## trim all chains to shortest mcmc chain, else 20000 samples
   if(exists('trait.mcmc')) {
@@ -72,7 +75,7 @@ for (i in seq(pft.names)){
     traits <- NA
     samples.num <- 20000
   }
-
+  
   priors <- rownames(prior.distns)
   for (prior in priors) {
     if (prior %in% traits) {
@@ -82,29 +85,39 @@ for (i in seq(pft.names)){
     }
     trait.samples[[pft.name]][[prior]] <- samples
   }
+}
 
-  ## subset the trait.samples to ensemble size using Halton sequence 
-  if('ensemble' %in% names(settings) && settings$ensemble$size > 0) {
-    ensemble.samples[[pft.name]] <- get.ensemble.samples(settings$ensemble$size, trait.samples[[pft.name]])
-    write.ensemble.configs(settings$pfts[[i]], ensemble.samples[[pft.name]], 
-                           host, outdir, settings)
-  }
-  
+## NEED TO IMPLEMENT:
+## Load Environmental Priors and Posteriors
 
-  if('sensitivity.analysis' %in% names(settings)) {
-    if( is.null(settings$sensitivity.analysis)) {
-      print(paste('sensitivity analysis settings are NULL'))
-    } else {
-      quantiles <- get.quantiles(settings$sensitivity.analysis$quantiles)
-      sa.samples[[pft.name]] <-  get.sa.samples(trait.samples[[pft.name]], quantiles)
-      write.sa.configs(settings$pfts[[i]], sa.samples[[pft.name]], 
-                       host, outdir, settings)
-    }
+
+## write SENSITIVITY ANALYSIS
+if('sensitivity.analysis' %in% names(settings)) {
+  quantiles <- get.quantiles(settings$sensitivity.analysis$quantiles)
+  if( is.null(settings$sensitivity.analysis)) {
+    print(paste('sensitivity analysis settings are NULL'))
+  } else {
+    sa.samples <-  get.sa.sample.list(trait.samples, 
+        env.samples,
+        quantiles)
+    write.sa.configs(settings$pfts, sa.samples, 
+        host, outdir, settings)
   }
 }
 
+## Write ENSEMBLE
+if('ensemble' %in% names(settings) && settings$ensemble$size > 0) {
+  ## subset the trait.samples to ensemble size using Halton sequence 
+  ensemble.samples <- get.ensemble.samples(settings$ensemble$size, trait.samples, env.samples)
+  write.ensemble.configs(settings$pfts, ensemble.samples, 
+      host, outdir, settings)
+}
+
+
+
+
 save(ensemble.samples, trait.samples, sa.samples, settings,
-     file = paste(outdir, 'samples.Rdata', sep=''))
+    file = paste(outdir, 'samples.Rdata', sep=''))
 
 ## Make outdirectory, send samples to outdir
 
@@ -112,12 +125,12 @@ if(host$name == 'localhost'){
   if(!host$outdir == outdir) {
     dir.create(host$outdir)
     file.copy(from = paste(outdir, 'samples.Rdata', sep=''),
-              to   = paste(host$outdir, 'samples.Rdata', sep = ''),
-              overwrite = TRUE)
+        to   = paste(host$outdir, 'samples.Rdata', sep = ''),
+        overwrite = TRUE)
   }
 } else {  
   mkdir.cmd <- paste("'if ! ls ", host$outdir, " > /dev/null ; then mkdir -p ", host$outdir," ; fi'",sep='')
   system(paste("ssh", host$name, mkdir.cmd))
   system(paste('rsync -routi ', paste(outdir, 'samples.Rdata', sep=''),
-               paste(host$name, ':', host$outdir, sep='')))
+          paste(host$name, ':', host$outdir, sep='')))
 }
